Skip using fetch unnecessarily to load our Base64 wasm
It's quick & convenient given a data URI, but it's also very easy to
just get base64, sychronously decode it, and load it (and
instantiateStreaming has little benefit imo when we've already loaded
the entire data as a string upfront anyway).

This simplifies nicely, but also adds support for Node environments
(<v18) where Fetch isn't available globally.