Copy data during output to fix apparent race condition
Documented in https://github.com/SteveSanderson/xzwasm/pull/1. We can't
use type:bytes AFAICT, as bytes returned from WASM are not
transferrable (the memory is still managed by WASM?) and so need to be
cloned regardless.

For our use cases, a little memory cloning isn't a big deal so this is
fine.